move ORTHOGRAPHY_NAME to vars.py
