+++ CreeDictionary/utils/vars.py
@@ -1,1 +1,6 @@
 DEFAULT_ORTHOGRAPHY = "Latn"
+ORTHOGRAPHY_NAME = {
+    "Latn": "SRO (êîôâ)",
+    "Latn-x-macron": "SRO (ēīōā)",
+    "Cans": "Syllabics",
+}
